# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.AccessRuleResponse:
    properties:
      role:
        type: string
      subjectId:
        type: string
      subjectType:
        type: string
    type: object
  api.CreateWorkspaceRequestPayload:
    properties:
      executionProfile:
        additionalProperties: true
        type: object
      owner:
        type: string
      ref:
        type: string
      repo:
        type: string
    type: object
  api.GrantAccessRequestPayload:
    properties:
      role:
        type: string
      subjectId:
        type: string
      subjectType:
        type: string
    type: object
  api.RemoveAccessRequestPayload:
    properties:
      subjectId:
        type: string
      subjectType:
        type: string
    type: object
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Workspace:
    properties:
      actualState:
        type: string
      createdAt:
        type: string
      desiredState:
        type: string
      executionProfile:
        items:
          type: integer
        type: array
      id:
        type: string
      owner:
        type: string
      ref:
        type: string
      repo:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Stateless Control Plane API for QuickSpaces.
  title: QuickSpaces Control Plane API
  version: "1.0"
paths:
  /access:
    delete:
      consumes:
      - application/json
      description: Removes a global GitHub access rule.
      parameters:
      - description: Remove access request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RemoveAccessRequestPayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Remove an access rule
      tags:
      - Access
    get:
      description: Returns all configured GitHub access rules.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AccessRuleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: List access rules
      tags:
      - Access
    post:
      consumes:
      - application/json
      description: Inserts or updates a global GitHub access rule.
      parameters:
      - description: Grant access request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GrantAccessRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AccessRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Grant or update an access rule
      tags:
      - Access
  /auth/login:
    get:
      description: Redirects a client to GitHub to begin OAuth authentication.
      produces:
      - application/json
      responses:
        "302":
          description: Found
      summary: Redirect user to GitHub OAuth login
      tags:
      - Auth
  /auth/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Exchanges an OAuth2 code for an access token (used by client-side
        apps).
      parameters:
      - description: OAuth authorization code
        in: formData
        name: code
        required: true
        type: string
      - description: PKCE code verifier (if PKCE was used)
        in: formData
        name: code_verifier
        type: string
      - description: Grant type (authorization_code)
        in: formData
        name: grant_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Exchange OAuth code for token
      tags:
      - Auth
  /health:
    get:
      description: Returns whether the control plane API is healthy.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - Health
  /workspaces:
    get:
      description: Returns all configured workspaces.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Workspace'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: List workspaces
      tags:
      - Workspaces
    post:
      consumes:
      - application/json
      description: Creates a workspace with a desired state and execution profile.
      parameters:
      - description: Create workspace request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateWorkspaceRequestPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Create a new workspace
      tags:
      - Workspaces
  /workspaces/{id}:
    get:
      description: Returns the workspace identified by its ID.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Workspace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get a workspace
      tags:
      - Workspaces
  /workspaces/{id}/reconcile:
    post:
      description: Reconciles the configured workspace state with the execution environment.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Workspace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Reconcile a workspace
      tags:
      - Workspaces
  /workspaces/{id}/start:
    post:
      description: Transitions the workspace to the running state.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Workspace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Start a workspace
      tags:
      - Workspaces
  /workspaces/{id}/stop:
    post:
      description: Transitions the workspace to the stopped state.
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Workspace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Stop a workspace
      tags:
      - Workspaces
schemes:
- http
swagger: "2.0"